# web/openapi.yaml
openapi: 3.0.3
info:
  title: Xray Checker API
  description: API for monitoring Xray proxy status and configuration
  version: "1.0"
servers:
  - url: /
security:
  - basicAuth: []
paths:
...
  /api/v1/proxies/{stableID}:
    get:
      summary: Get proxy by ID
      description: Returns information for a specific proxy. Configuration details are included only when WEB_SHOW_DETAILS=true.
      tags:
        - Proxies
      parameters:
        - name: stableID
          in: path
          required: true
          schema:
            type: string
          description: Proxy Stable ID (16-character hash)
      responses:
        '200':
          description: Proxy details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProxyInfo'
        '404':
          description: Proxy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorResponse'
...
  /api/v1/system/ip:
    get:
      summary: Get current IP
      description: Returns the current detected IP address
      tags:
        - System
      responses:
        '200':
          description: Current IP address
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SystemIPResponse'
        '500':
          description: Failed to get IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorResponse'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error message (only present on failure)
    APIErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Proxy not found"
...
    ProxyInfo:
      type: object
      properties:
        index:
          type: integer
          example: 0
        stableId:
          type: string
          example: "a1b2c3d4e5f67890"
        name:
          type: string
          example: "US-Server-1"
        subName:
          type: string
          description: Subscription name this proxy belongs to
          example: "Premium VPN"
        group:
          $ref: '#/components/schemas/ProxyGroupInfo'
        server:
          type: string
          example: "192.168.1.1"
        port:
          type: integer
          example: 443
        protocol:
          type: string
          enum: [vless, vmess, trojan, shadowsocks]
          example: "vless"
        proxyPort:
          type: integer
          example: 10000
        online:
          type: boolean
          example: true
        latencyMs:
          type: integer
          format: int64
          example: 150
        details:
          allOf:
            - $ref: '#/components/schemas/ProxyDetails'
          description: Optional configuration details, present only when WEB_SHOW_DETAILS=true.
    ProxyGroupInfo:
      type: object
      description: Metadata for JSON subscription balancer groups.
      properties:
        key:
          type: string
          description: Stable group key encoded as a JSON tuple of subscription name and group name.
          example: '["Premium VPN","Germany"]'
        name:
          type: string
          example: "Germany"
        index:
          type: integer
          description: Zero-based index inside the group.
          example: 0
        size:
          type: integer
          description: Number of proxies in the group.
          example: 3
        collapsed:
          type: boolean
          example: true
    ProxyCredentialInfo:
      type: object
      properties:
        type:
          type: string
          example: "uuid"
        value:
          type: string
          description: Masked credential value.
          example: "abcd...1234"
    ProxyInboundInfo:
      type: object
      properties:
        listen:
          type: string
          example: "127.0.0.1"
        port:
          type: integer
          example: 10000
        protocol:
          type: string
          example: "socks"
        tag:
          type: string
          example: "US-Server-1_vless_0_Inbound"
    ProxyOutboundInfo:
      type: object
      properties:
        tag:
          type: string
          example: "US-Server-1_0"
        protocol:
          type: string
          enum: [vless, vmess, trojan, shadowsocks]
          example: "vless"
        address:
          type: string
          example: "192.168.1.1"
        port:
          type: integer
          example: 443
    ProxyDetails:
      type: object
      properties:
        protocol:
          type: string
          enum: [vless, vmess, trojan, shadowsocks]
          example: "vless"
        address:
          type: string
          example: "192.168.1.1"
        port:
          type: integer
          example: 443
        transport:
          type: string
          example: "tcp"
        security:
          type: string
          example: "reality"
        sni:
          type: string
          example: "example.com"
        host:
          type: string
          example: "example.com"
        path:
          type: string
          example: "/ws"
        mode:
          type: string
          example: "auto"
        flow:
          type: string
          example: "xtls-rprx-vision"
        encryption:
          type: string
          example: "none"
        headerType:
          type: string
          example: "none"
        fingerprint:
          type: string
          example: "chrome"
        publicKey:
          type: string
          example: "public-key"
        shortId:
          type: string
          example: "abcd1234"
        serviceName:
          type: string
          example: "grpc-service"
        multiMode:
          type: boolean
          example: false
        allowInsecure:
          type: boolean
          example: false
        alpn:
          type: array
          items:
            type: string
          example: ["h2", "http/1.1"]
        credential:
          $ref: '#/components/schemas/ProxyCredentialInfo'
        inbound:
          $ref: '#/components/schemas/ProxyInboundInfo'
        outbound:
          $ref: '#/components/schemas/ProxyOutboundInfo'
        rawXhttpSettings:
          type: object
          additionalProperties: true
          example:
            host: "example.com"
            path: "/xhttp"
            mode: "auto"
...
    SystemIPResponse:
      type: object
      properties:
        ip:
          type: string
          example: "203.0.113.1"